feat: Git 이벤트 기반 인덱싱, 브랜치/레포 DB 동기화, JGit REST 인터페이스 도입 및 관련 구조 리팩토링

+++ src/main/kotlin/com/example/gitserver/module/gitindex/infrastructure/s3/S3BlobUploader.kt
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Component
 import software.amazon.awssdk.core.sync.RequestBody
 import software.amazon.awssdk.services.s3.S3Client
 import software.amazon.awssdk.services.s3.model.PutObjectRequest
+import software.amazon.awssdk.services.s3.model.HeadObjectRequest
 
 private val log = mu.KotlinLogging.logger {}
 
@@ -14,29 +15,42 @@ class S3BlobUploader(
     @Value("\${cloud.aws.s3.bucket}") private val bucketName: String
 ) {
     /**
-     * S3에 Blob 데이터를 업로드합니다.
-     * @param repositoryId 저장소 ID
+     * S3에 Blob을 "전역 해시" 기반으로 저장합니다.
+     * 이미 존재하는 blob은 중복 저장하지 않습니다.
      * @param hash Blob 해시
      * @param bytes Blob 데이터
-     * @return S3에 저장된 Blob의 키
+     * @return S3에 저장된 Blob의 key
      */
-    fun upload(repositoryId: Long, hash: String, bytes: ByteArray): String {
-        log.info { "업로드 시작 $repositoryId blob to $hash" }
-        val key = "blobs/$repositoryId/$hash"
+    fun upload(hash: String, bytes: ByteArray): String {
+        val key = "blobs/$hash"
 
-        val putRequest = PutObjectRequest.builder()
-            .bucket(bucketName)
-            .key(key)
-            .contentLength(bytes.size.toLong())
-            .build()
+        val alreadyExists = try {
+            s3Client.headObject(
+                HeadObjectRequest.builder()
+                    .bucket(bucketName)
+                    .key(key)
+                    .build()
+            )
+            log.info { "이미 존재하는 blob, 업로드 생략 - hash=$hash, key=$key" }
+            true
+        } catch (e: Exception) {
+            false
+        }
 
-        s3Client.putObject(
-            putRequest,
-            RequestBody.fromBytes(bytes)
-        )
+        if (!alreadyExists) {
+            val putRequest = PutObjectRequest.builder()
+                .bucket(bucketName)
+                .key(key)
+                .contentLength(bytes.size.toLong())
+                .build()
 
-        log.info { "블롭 업로드 S3 - repositoryId=$repositoryId, hash=$hash, key=$key" }
+            s3Client.putObject(
+                putRequest,
+                RequestBody.fromBytes(bytes)
+            )
+            log.info { "새로운 blob 업로드 완료 - hash=$hash, key=$key, size=${bytes.size} bytes" }
+        }
+
         return key
-
     }
 }
